+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/tests/fargate/test_02_clean.py
@@ -2,6 +2,10 @@ import unittest
 from unittest.mock import ANY, Mock, patch
 
 import projects._07_workforce_characteristics._01_starters_leavers_vacancies.fargate._02_clean as job
+from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
+    AscwdsWorkplaceCleanedColumns as AWPClean,
+)
+from utils.column_names.slv_job_role_columns import SLVJobRoleColumns as SLVCols
 
 PATCH_PATH = "projects._07_workforce_characteristics._01_starters_leavers_vacancies.fargate._02_clean"
 
@@ -11,12 +15,12 @@ class MainTests(unittest.TestCase):
     CLEANED_DATA_DESTINATION = "some/destination"
 
     @patch(f"{PATCH_PATH}.utils.sink_to_parquet")
-    @patch(f"{PATCH_PATH}.cUtils.deduplicate_slv_over_time")
+    @patch(f"{PATCH_PATH}.cUtils.remove_repeated_values_over_time")
     @patch(f"{PATCH_PATH}.utils.scan_parquet")
     def test_main_runs(
         self,
         scan_parquet_mock: Mock,
-        deduplicate_slv_over_time_mock: Mock,
+        remove_repeated_values_over_time_mock: Mock,
         sink_to_parquet_mock: Mock,
     ):
         job.main(
@@ -25,8 +29,11 @@ class MainTests(unittest.TestCase):
         )
 
         scan_parquet_mock.assert_called_once_with(self.MERGED_DATA_SOURCE)
-        deduplicate_slv_over_time_mock.assert_called_once_with(
-            scan_parquet_mock.return_value
+        remove_repeated_values_over_time_mock.assert_called_once_with(
+            scan_parquet_mock.return_value,
+            columns_to_clean=[SLVCols.starters, SLVCols.leavers, SLVCols.vacancies],
+            partition_by_column=AWPClean.establishment_id,
+            date_column=AWPClean.ascwds_workplace_import_date,
         )
 
         sink_to_parquet_mock.assert_called_once_with(